Moved the convert() method to the presenter class
